requirements.txt
pyautogui
pyperclip
pygetwindow
keyboard
websocket-client
requests
